+++ LibOS/shim/src/elf/shim_rtld.c
@@ -356,7 +356,7 @@ __map_elf_object (struct shim_handle * file,
     if (file && (!file->fs || !file->fs->fs_ops))
         return NULL;
 
-    int (*read) (struct shim_handle *, void *, size_t) =
+    int64_t (*read) (struct shim_handle *, void *, size_t) =
         file ? file->fs->fs_ops->read : NULL;
     int (*mmap) (struct shim_handle *, void **, size_t, int, int, off_t) =
         file ? file->fs->fs_ops->mmap : NULL;
@@ -846,7 +846,7 @@ int free_elf_object (struct shim_handle * file)
     return 0;
 }
 
-static int __check_elf_header (void * fbp, int len)
+static int __check_elf_header (void * fbp, size_t len)
 {
     const char * errstring __attribute__((unused));
 
@@ -944,7 +944,7 @@ static int __read_elf_header (struct shim_handle * file, void * fbp)
     if (!file->fs || !file->fs->fs_ops)
         return -EACCES;
 
-    int (*read) (struct shim_handle *, void *, size_t) =
+    int64_t (*read) (struct shim_handle *, void *, size_t) =
         file->fs->fs_ops->read;
     int (*seek) (struct shim_handle *, off_t, int) =
         file->fs->fs_ops->seek;
